blocked check

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -17,7 +17,6 @@ class CategoryController extends Controller
     {
         $user = Auth::user();
         if ($user->is_blocked) {
-            // return "you are blocked";
             return view('blocked.index');
         } else {
             $data = DB::table('categories as c')
@@ -43,16 +42,21 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
-        $data['name'] = $request->name;
+        $user = Auth::user();
+        if ($user->is_blocked) {
+            return view('blocked.index');
+        } else {
+            $data['name'] = $request->name;
 
-        $check = Category::where('name', $request->name)->first();
-        if ($check) {
-            return redirect()->route("admin.categories.index")->with('danger', '"' . $check->name . '" already exist.');
+            $check = Category::where('name', $request->name)->first();
+            if ($check) {
+                return redirect()->route("admin.categories.index")->with('danger', '"' . $check->name . '" already exist.');
+            }
+
+            $data['created_at'] = date('Y-m-d H:i:s');
+            Category::create($data);
+            return redirect()->route("admin.categories.index")->with('success', 'category has been created successfully.');
         }
-
-        $data['created_at'] = date('Y-m-d H:i:s');
-        Category::create($data);
-        return redirect()->route("admin.categories.index")->with('success', 'category has been created successfully.');
     }
 
     /**
@@ -68,8 +72,13 @@ class CategoryController extends Controller
      */
     public function edit(string $id)
     {
-        $data = Category::select('*')->find($id);
-        return view("admin.category.editcategories", ['data' => $data]);
+        $user = Auth::user();
+        if ($user->is_blocked) {
+            return view('blocked.index');
+        } else {
+            $data = Category::select('*')->find($id);
+            return view("admin.category.editcategories", ['data' => $data]);
+        }
     }
 
     /**
@@ -77,17 +86,22 @@ class CategoryController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $data['name'] = $request->name;
+        $user = Auth::user();
+        if ($user->is_blocked) {
+            return view('blocked.index');
+        } else {
+            $data['name'] = $request->name;
 
-        $check = Category::where('name', $request->name)->whereNotIn('id', [$id])->first();
-        if ($check) {
-            return redirect()->route("admin.categories.edit", $id)->with('danger', '"' . $check->name . '" already exist.');
+            $check = Category::where('name', $request->name)->whereNotIn('id', [$id])->first();
+            if ($check) {
+                return redirect()->route("admin.categories.edit", $id)->with('danger', '"' . $check->name . '" already exist.');
+            }
+
+            $data['name'] = $request->name;
+            $data['updated_at'] = date('Y-m-d H:i:s');
+            Category::where('id', $id)->update($data);
+            return redirect()->route("admin.categories.index")->with('success', 'category has been edited successfully.');
         }
-
-        $data['name'] = $request->name;
-        $data['updated_at'] = date('Y-m-d H:i:s');
-        Category::where('id', $id)->update($data);
-        return redirect()->route("admin.categories.index")->with('success', 'category has been edited successfully.');
     }
 
     /**
@@ -95,7 +109,12 @@ class CategoryController extends Controller
      */
     public function destroy(string $id)
     {
-        Category::where('id', $id)->delete();
-        return redirect()->route('admin.categories.index')->with('success', 'category has been deleted successfully.');
+        $user = Auth::user();
+        if ($user->is_blocked) {
+            return view('blocked.index');
+        } else {
+            Category::where('id', $id)->delete();
+            return redirect()->route('admin.categories.index')->with('success', 'category has been deleted successfully.');
+        }
     }
 }
